[project @ unified interface from polynomials to Singular] Removed the MPolynomial[Ring]_singular_repr classes from SAGE and replaced them by Polynomial[Ring]_singular_repr classes which are base classes of MPolynomial_polydict_domain, Polynomial_field, and Polynomial_mod_p. These new Polynomial[Ring]_singular_repr classes unify conversions to Singular and allow e.g. to have a MPolynomial_polydict to be coercable to Singular and MacCaulay2 as well. The MacCaulay2 part of this API change is not implemented yet.
No checks are performed in MPolynomial_* classes if conversion to Singular is supported but they
depend on their respective MPolynomialRing_* classes. (So there is some kind of usability issue
as the error message that a ring couldn't be coerced might not be clear to the user). These ring
classes have a new attribute _has_singular which must be True to allow conversion. Extending
Singular conversion to e.g. Reals should now only require changes to polynomial_singular_interface.py.

'make test' passes.