#5120: An alternative implementation of the Unique Representation design pattern
Class deriving from sage.structure.UniqueRepresentation inherit a
unique representation behavior for their instances (and consistent
default implementations of equality testing, copying, pickling, ...).

See the documentation for a brief discussion of the relative merits of
UniqueRepresentation? and UniqueFactory?.

As a prerequisite, this patch implements sage.misc.ClasscallMetaclass,
a (trivial) metaclass for customizing class calls via a static method
of the class.

This class is used extensively in upcoming sage-combinat patches:
 - #5879 (crystals)
 - #5891 categories
 - free modules
 - automatic monoid