#5991 Add a standard constructor for dynamic classes
This patch implements sage.structure.dynamic_class.dynamic_class, for
constructing dynamically new python classes. The constructed classes
can be pickled, and have unique representation.

The patch includes a discussion on the relevance of dynamic classes for Sage.

Depends on #5985 for the pickling and #5120

Will be used by the category framework #5891, (and sage-words)

Includes a small patch to sage/misc/sageinspect so that class with a metaclass
are *not* considered as class instances.

Issue: is sage.structure.dynamic_class.dynamic_class the natural location for this?