consolidate in plotting all extraction of variables, ranges, and fast_float setup
Currently, code for extracting variables, dealing with ranges of variables, and making the functions fast_float is scattered throughout the plotting directory. There are multiple implementations, each having its own quirks.

This patch consolidates all of this to two functions in sage.plot.misc and makes all the necessary changes to use this consolidated function.