Trac 9808 - fix Scipy not compiling properly on OS X 10.6 because of Numpy
A change in how Numpy detects the arch flags on OS X is leading to too many binaries being included in fat binaries.  This patch does the minimal change, simply removing the ppc64 arch flag on Snow Leopard.

When Numpy 1.5.1 is released, check to see if this patch is still necessary.