#5847 (ecm-6.3.p2): Don't unconditionally add '-fPIC'; remove code from testing ("if false...").
Only add '-fPIC' to CFLAGS if we're not building a shared library and '--with-pic'
was not given. (libtool correctly builds the shared library without that.)

Not using PIC is usually faster, so we don't build the *static* library with PIC
in case the Sage library uses a shared one (which of course uses PIC).