trac #7191 dirichlet characters over CC
Changes to dirichlet characters to support CC as a base field. In addition
to changing the discrete log table lookup, this makes changes to
gauss_sum(), gauss_sum_numerical(), is_even(), and is_odd(). It is
possible that further changes have to be made, and maybe some more
doctests need to be added.