Trac #12780: Remove all domain assignments in simplification functions.
This removes assignments to Maxima's 'domain' variable in the
simplify_radical() and simplify_log() functions. This setting should
be up to the user ultimately; but for now, should at least be
consistent throughout our symbolics.

We also fix expand_log(), which used to assume that the domain was
'complex' when it was called. Now, it makes note of the current domain
when it was called and restores it before returning.