Trac #14252: Implementation of quantum LS paths
This patch cleans up the interface of CrystalOfLSPaths. In particular, the input is now a weight, whereas before it took a cartan type and a list of coefficients. The class call function still allows this input format.

In addition, a new implementation of one-column Kirillov-Reshetikhin crystals is given in terms of level-zero CrystalOfLSPaths together with the energy function and one-dimensional configuration sum.