Use \cdot to separate coefficient from following digit in latex mul output.
Fixes #18. This is Sage ticket #13356.

The overall coefficient is printed first in the output of mul objects. In the
latex representation, if the first character of the rest of the output is a
digit, this may produce confusing results.

sage: e = 2 * 2^(1/3)
sage: print latex(e)
2 \, 2^{\left(\frac{1}{3}\right)}

Now we use \cdot as a separator in this case to get:

sage: latex(2 * 2^x)
2 \cdot 2^{x}