Trac #9556: add dynamic methods to symbolic expressions
If a symbolic function represents the evaluated form of a symbolic function,
it is useful in some cases to extract properties of this specific instance of
the function. The methods used for this purpose should be defined close to the
body of the symbolic function, but available as methods of symbolic
expressions.

This patch uses dynamic class magic to take methods defined in a subclass
named EvaluationMethods in the function body and make them available as
methods of symbolic expressions whenever the expression is the evaluated form
of that function.

This particular version of the implementation is built on ideas from Simon
King, Mike Hansen and Volker Braun.

One use case for this feature is #2516.