Trac #15023: implements correct default branching rules
The existing default branching rule in weyl_characters.py is only correct in
certain cases. No error is returned if the branching rule is not specified,
even if no branching rule is known to Sage. This patch corrects the
situation by implementing default branching rules in a few common, obvious
cases, and raising a Value error if no default branching rule is known.