Make sure we do not touch files outside of sage in the dev scripts.
Some git commands (most prominently git clean) are very dangerous if the user
is not inside of the sage directory. Here we make sure that at least the dev
scripts are always called from within sage. If dev.git is used directly, the
user can still wipe its home. This seems to be ok, because dev.git is just like
!git.