import_patch() used non-existing parameter quiet for reset_to_clean_working_directory()
David Roe had seen lots of "not deleting directory ..." messages when running
this command. This seems to be a problem in your git config or the version of
git you are using. I don't see these messages. Feel free to add this again if
the problem shows up for too many people.