Add quotient with remainder and euclidean degree tests to the category of euclidean domains
For such test methods, there has to be a canonical name for the euclidean
degree. Just calling it degree() is very confusing: in the context of Integers
or p-adics, this term does not make too much sense. Additionally, there could
already be a degree() for some elements which happens to be different from the
degree function of the euclidean domain.
Note, that I could not just set euclidean_degree=degree for polynomials because
polynomials (even with coefficients in a field) can have a parent where this is
not the euclidean degree (e.g. in a field K[x]/(f)).