fix output of enumerate_totallyreal_*
The output of enumerate_totallyreal_fields_prim,
enumerate_totallyreal_fields_all, and enumerate_totallyreal_fields_rel
are in general simply the output of Pari. They should belong to
Sage proper, so that the output can be later used right away.

This patch makes sure that the output of integers are Integers,
and the output that looks like polynomials belong to PolynomialRing