Use simple coefficients if base is multivariate polynomial.
If the base ring is a multivariate polynomial ring, then the discriminant is
now computed in two steps.  First all non-zero coefficients of the
polynomial are replaced with variables, and the discriminant is computed
using these variables.  Then the original coefficients are substituted into
the result of this computation.  This should be faster in many cases,
particularly with big coefficients.  In those cases, the matrix computation
on simple generators is a lot faster, and the final substitution is cheaper
than carrying all that information along at all times.