mpmath integration for builtin functions.
This is an import of the patch I previously attached to the trac ticket.
I'm applying it against the version fo sage against which it apparently was
created, namely version 5.10.

On the one hand, this patch addressed the original issue reported in this
ticket: it makes exp(a) work if a is an mpmath object.  It does so by not
only checking whether an object has an attribute of a given name, but also
whether that attribute is actually callable.

On the other hand, this re-routes several standard functions to their mpmath
implementation if one of the arguments is an mpmath object.  This is done in
such a way as to avoid importing mpmath until it is actually needed.  One
drawback here is that we determine when it will be needed based on the
module name.  So subclasses of mpmath types which belong to some other
module will not trigger the same behavior.  Experiments indicate that the
performance hit of this modification seems to be neglibile.

As a drive-by fix, a number of call-or-catch-AttributeError sequences were
modified to try-catch-else, to make sure that a missing method triggers the
exception handler but that an error invoking the method will get propagated.