Implemented NumberFieldStructure classes to improve caching of (absolute) number fields.
The structure() of a number field typically returns morphisms from and to a
number field. The number field is required as the domain/codomain of these
morphism so these structure morphisms can not be passed in as parameters (e.g.
to a caching factory) when the number field is created. A NumberFieldStructure
wraps the parameters which are required to create these morphisms and only
creates the morphisms when they are needed (after the number field has been
created).