Introduce a caching function universal_discriminant.
This approach of using a separate function to compute universal
discriminants was suggested by both John Cremona and Peter Bruin.

This implementation does not call Polynomial.discriminant, which has the
benefit that it may be called by that method without causing an infinite
loop and without need for an extra argument for algorithm selection.
Another benefit is that the special case for zero divisors can be avoided,
as can be the polynomial quotient field.

I chose to implement this function in the polynomial_element module, but
John suggested placing it in invariant_theory, so I'd like to document this
alternative suggestion here.