Build maxima fasl without asdf.
Instead of some ecl scripting in the package installation script, this
approach patches maxima.system in order to call c::build-fasl similar to the
existing c::build-program call. This approach has been used on Gentoo for
some time now, and it works for newer versions of ECL and Maxima as well.