Explanations for >1 final components, periodic final components
Previously, for transducers with more than one final component, a
NotImplementedError was raised trying to give some short
explanation. This explanation is now moved to the doctest testing for
this case, and the explanation has been extended to show what the
result would be in this specific case.

A similar explanation is now given for the case of aperiodic final
components