Ticket #16187 reviewer patch: Beautify output of sage --{optional,experimental}
Don't list *all* installed packages (especially since the heading is
"optional/experimental packages"), but instead only the optional /
experimental Sage packages available on the server, and *next to* each
package whether it is already installed (same version), not installed,
or in case a different version of the package is installed, the latter.

Note that for some packages there is for example a standard version, and
a different optional or experimental one available.  One may also have an
outdated version of e.g. an optional spkg already installed.

Packages without a (proper) version (following the first dash in their name)
are also handled correctly (I think).

Minor clean up (and change to usual indentation).