FiniteStateMachine._latex_: typeset final_word_out, introduce relevant options
If a state has non-empty .final_word_out, this final output word is typeset
as a label of a leaving edge.

On the way, support options for tikz' accepting/.style=accepting by arrow,
accepting distance and accepting where.

As long as #19161 is not implemented, transitional code makes sure that
all final states have an attribute final_word_out.