FiniteStateMachine: Convenience Formatting functions for digit problems
When using negative digits, it is quite customary to write e.g. \overline{1}
instead of -1 in LaTeX. Introduce FiniteStateMachine.format_letter_negative
for this purpose.

Furthermore, In digit expansions, digits are frequently processed from
the least significant to the most significant position, but it is
customary to write the least significant digit at the right-most
position. Therefore, the labels have to be reversed. Provide a method
FiniteStateMachine.format_transition_label_reversed for this purpose.