FiniteStateMachine._latex_: typeset state labels in math mode (#16200)
Currently, state labels are not typeset in LaTeX's math mode, but
transition labels are. Furthermore, the default behaviour is to pipe
the label through latex(), which by definition produces output
suitable for math mode. Therefore, I consider this to be a bug and
change the behaviour.

This is an incompatible change; however, relying on the old behaviour
will either make no difference (if the label was an integer or a
string (this is protected by \text anyway)) or lead to an error when
typesetting with tikz.

While doing that, I also include a space after the comma in the
coordinate tuples (PEP 8 probably does not apply to TeX code included
in a docstring, but nevertheless ...).