Trac 10963: suppressed now useless category example MyGroupAlgebra + tiny doc improvement
With the merging of the former category GroupAlgebras with the
functorial construction Groups.Algebras, the example
sage.categories.examples.group_algebras.MyGroupAlgebra has become
completely useless. Since this example was not meant to be used in
real code and was marked as dubious, we don't need to go through a
deprecation phase. Suppressed.