FiniteStateMachine.with_final_word_out: New method
Constructs a finite state machine with final output words for all
states by implicitly reading trailing letters until a final state is
reached.

This is e.g. useful for finite state machines operating on digit
expansions: there, it is sometimes required to read a sufficient
number of trailing zeros (at the most significant positions) in order
to reach a final state and to flush all carries. In this case, this
method constructs an essentially equivalent finite state machine in
the sense that it not longer requires adding sufficiently many
trailing zeros. However, it is the responsibility of the user to make
sure that if adding trailing zeros to the input anyway, the output is
equivalent.