Fixed resultant for polynomials over function fields.
Polynomials over extensions of rational function fields can not be converted to
PARI. Now an appropriate exception is raised.
Also, if the order of the variables is not like PARI expects it, then a
PariError is raised. If this happens, rely on the naive implementiation, i.e.,
compute the determinant of the Sylvester matrix.