Improved LocalGeneric._test_residue()
If there is not enough precision/the element has negative valuation, we can not
compute the residue. An exception should be raised in that case. In the first
case this should be PrecisionError. In the other case, a ValueError seems to
make sense though we might change this later to something more specific.