finite_state_machine.FSMState: New method fully_equal
finite_state_machine.FSMState.__eq__ only compares labels of states,
but not their other attributes such as is_initial, is_final, final_word_out,
word_out, color.

The reason is that states have to be hashable and that their hash only depends on the label --- therefore, states with the same labels have to be equal.

This is no problem as long as states of the same finite state machine are compared. When comparing two finite state machines, this distinction matters.

Therefore, a new method FSMState.fully_equal is introduced and used in
FiniteStateMachine.__eq__ .