Include dot in extension for graphics_filename().
This makes the function more similar to tmp_filename(), so people may
have less trouble remembering where to include the dot and where not to.
This also enables the use of suffixes which don't start in a dot, like
those starting in a dash which are used by Graphics3d.show().

The old behaviour is deprecated but kept, at least for extensions starting
in neither a dot nor a dash.  That compatibility code may be removed a year
after this change here got released.