Generate names for 3d graphics files independently.
Instead of basing everything on a single call to graphics_filename,
we now have one call for (almost) every file which is being generated.
This ensures that the files with the corresponding extensions are indeed new.
Up until now, creating multiple 3d results in a single cell would likely
lead to clashes.