trac #16580: use OrderedDict instead of dict to have non-random output
In http://trac.sagemath.org/ticket/16580#comment:3 it turned out that
the order of the output of the transitions in _latex_ seems to be
architecture-dependent. While this has no relevance for the actual
latex output, it is annoying for doctests.

Therefore, we now use an OrderedDict instead of a dict in order to
output the transitions ordered by the labels of the states.

Consequently, the order of some transitions in four doctests had
to be permuted to fit to the new output.