Minor change in change_ring function for scheme morphisms.
Changed for loop to list comprehension structure for compactness. The loop iterates over the defining polynomials of the morphism. Instead of trying to compute the length of the list, the upper bound was just replaced with the list length function. This addresses the issue of the upper bound not being appropriate for affine morphisms.