Brandt modules: extend _ideal_products() method
Some computations, e.g. monodromy_weights() requires only the
diagonal of the brandt matrices. It is convenient then to extend the
method _ideal_products() adding an option "diagonal_only" (defaults
False) to compute only the diagonal elements.

This is important for large dimensions where one doesn't use the brandt
matrix method to compute hecke operators, but rather the 'direct' method
which doesn't require to compute the full O(n^2) products.

It is convenient to include this option in the same method so that the
caching of the ideals (and so of the theta series) is shared.