animate: Pass keyword arguments to plot constructor.
Some keyword arguments are only effective if passed to the plot.plot factory
function.  Passing them when saving the image doesn't have the same effect.
One example is xmin and xmax, which will affect the plot range as well.
So this change here should avoid a regression we've been observing.