Trac #12162: Add a simplify_full() method to the Matrix_symbolic_dense class.
Symbolic matrices can already be simplified with the less-powerful
simplify(), simplify_trig(), and simplify_rational() methods. But so
far a componentwise simplify_full() has not been available.

This patch adds a new method with a test case based on
Expression.simplify_full(). The implementation is a straightforward
list comprehension followed by a conversion back to the appropriate
matrix space.