Trac #17393: Add a domain warning to Expression.simplify_log().
The Expression.simplify_log() method should only be used on
expressions with real variables. This is implicit in the description,
but it deserves an explicit warning as well. This commit adds a
warning with an (counter)example over the complex numbers.

A few other minor improvements were made to the documentation:

  * Variables 'a', 'b', 'c', and 'd' are used instead of 'a1', 'a2',
    'b1', and 'b2' which could be misinterpreted given that they
    appeared in "math mode."

  * A test case that appeared under EXAMPLES was moved to TESTS.

  * The DETAIL section was mostly removed, since it was at best
    redundant but mostly irrelevant.

  * The underlying Maxima call was moved to an ALGORITHM section.

  * Small grammar and quoting cleanups.