Trac #17389: Remove unsafe simplifications from test_relation_maxima().
When testing equality with assumptions, one branch of
test_relation_maxima() in sage.symbolic.relation will try to simplify
a difference (left- minus right-hand side) to zero. But some of the
simplifications it uses, namely canonicalize_radical() and
simplify_log(), are unsafe for complex variables.

This commit removes those operations from the list, and adds two new
(safe) simplifications: simplify_factorial() and
simplify_rectform(). Test cases are added for bugs involving complex
variables, and for expressions that can be reduced to zero using the
new simplifications.