Doctests for each method of APngAssembler using tracing fixture.
While I personally believe that there is not much to be gained from
doctesting every tiny internal method, this lack of doctests is currently
preventing a positive review, so here they are.

I implemented doctests in the way that seemed the most reasonable to me:
show the role of each method in a realistic run of the code, using a tracing
fixture to exhibit the interaction of a given method with its caller (who
passes its arguments and may receive some return value), the object state
(via read and write access to attributes) and other methods (which are
called during the execution of a given method).