Trac #17556: Remove simplify_log() from simplify_full().
The simplify_log() method for symbolic expressions is only valid when
its argument (the expression) is real. There is an elementary example
-- now a doctest -- where simplify_log() produces an incorrect
result for complex expressions. Since Sage assumes that expressions
are complex by default, it's not safe to have simplify_log() as part
of simplify_full(); i.e. under the "simplify" moniker.

This commit removes simplify_log() from simplify_full(), and adds a
doctest for the new behavior. Since the 'one' algorithm was used in
simplify_log() within simplify_full(), the simplifications that it
performed were extremely neutered. Its removal is therefore no great
loss.