Trac #17556: Add simplify_rectform() to simplify_full().
There is a newer simplification, Expression.simplify_rectform(), which
is not performed as part of Expression.simplify_full(). The rectform
simplification is safe for complex expressions, so it makes sense to
include it as part of the "full" simplification list.

This commit also removes a redundant simplify_rational() left over
from the removal of simplify_log() from simplify_full().