Trac #17556: Add Expression.simplify_log() to Expression.simplify_real().
The Expression.simplify_log() simplification is only valid over the
real numbers. It has been removed from Expression.simplify_full() for
that reason, but there's no reason it can't be performed as part of
simplify_real() where it is quite clear that the expression must be
real.

This commit performs simplify_log() as part of simplify_real(), and
updates the documentation of simplify_real().