Add reproducible_repr for list and dict.
The special handling of lists is needed in order to perform recursive calls
of this function for each element of a list.

Also reduce reliance on comparison operators, by comparing representations
instead of (or at least in preference to) actual values. This has the
drawback that we don't sort numerically, which might seem strange.
But it has the benefit that we can establish a total order
even for containers of mixed and mutually not comparable types,
as long as distinct elements yield distinct representations.