Implement symbolic (radical) roots of polynomials.
This is considered a preliminary implementation, to be extended once #17516
gets implemented.  Other code which employs similar functionality is
encouraged to delegate to this method, so that they will benefit from
subsequent improvements here.  Note that the case where K is SR is unlikely
to benefit from Galois theory, so the core of the code will likely stay
around, as one of several cases and most likely as the default fallback.

Instead of expr=self(var), one might also have considered SR(self).  Tests
show that the former is faster (for a moderate rational polynomial), and the
latter fails altogether if K is SR.  So this substitution was chosen.

Choosing a fixed, long and unlikely-to-clash variable was inspired by the
appoach taken for the charpoly computation in #14403, suggested by nbruin.