Better inversion of module actions for correct division coercions.
In particular, if we have R1 -> R2 with R2 acting on M, then
the inverse action attempts to invert elements of R1 after coercing
to R2 rather than before.  This is particularly important when
the coercion map does not commute with taking the fraction field,
e.g. R1 = ZZ and R2 = GF(p).