Trac #17218: Change signature of acb_to_ComplexIntervalFieldElement
Previously, acb_to_ComplexIntervalFieldElement was given a
ComplexIntervalField and created the corresponding
ComplexIntervalFieldElement on its own.

This might be inefficient when called frequently, as the Python class
ComplexIntervalField has to be involved.

Now, a ComplexIntervalFieldElement can be created by
ComplexIntervalFieldElement._new by the caller, so this low-level
function avoids unnecessary overhead.