Fair_Cake_Cutting_game
The problem involves a heterogenous resource, such as a cake with different toppings, that is assumed to be divisible – it is possible to cut arbitrarily small pieces of it without destroying their value. The resource has to be divided among several partners who have different preferences over different parts of the cake, i.e., some people prefer the chocolate toppings, some prefer the cherries, some just want as large a piece as possible, etc. The division should be subjectively fair, i.e., each person should receive a piece that he/she believes to be a fair share.