Kuhn_poker
Kuhn poker is a zero sum two player game. The deck includes only three cards. One card is dealt to each player, then the first player must bet or pass, then the second player may bet or pass. If any player chooses to bet the opposing player must bet as well in order to stay in the round. After both players pass or bet the player with the highest card wins the pot.