Don't use file name argument to show method
Since 7747261d04e4ff0105fb514a2109b2d014396e2a removed the file name
argument to the show method of GraphicsArray (without deprecatuion
procedure), we should not document it being called in this way.

Instead we can save it to a file (which won't get displayed automatically)
to demonstrate the core idea of specifying a custom file name here.