Improved unicode delimiters for matrices, allow choosing the shape
Now we have a shape keyword argument, and the default is round for unicode.
New arguments are enforced to be keyword-only.
Extra tests ensure correct output for single row matrices.
Unicode literals are now all in one place, the rest are variables.
This avoids replacing in strings or splitting strings again.
Instead of iterating over the divs arrays, we now have a second array
containing the number of divs for each position, for O(1) access to that.
Some other minor things were cleaned up or simplified as well.