TeX rendering for unicode big delimiters
This prevents LaTeX from balking at the unicode symbols which represent
(parts of) big delimiters.  The symbols are rendered using their equivalent
from the computer modern math extensions font.  So the symbols themselves
look right, even though the spacing is very wrong.  Nevertheless, this
should convey the idea of what gets printed, and should suffice until we
switch to XeTeX or luaTeX with hopefully better Unicode support.

The codes for the characters were found in texmext.enc.  The corresponding
character names were found in the PostScript output of a small test file,
along with CMEX10 as the font, indicating the use of \fontfamily{cmex}.  The
\fontencoding{OMX} to go with that was found in amsmath.sty.  The argument
to \raisebox was a bit of trial and error, and pretty much a rough hack to
at least line up some of the delimiters with their corresponding matrix
rows.

Turning U+253C into + makes it fit in with the surrounding - and | used for
the horizontal and vertical lines.