Trac #18392: Add is_solid() and is_proper() methods to polyhedral cones.
The notion of a proper (closed, convex, solid, and pointed) cone is
useful in optimization because such cones give rise to partial orders
on the ambient vector space.

In sage, our polyhedral cones are closed and convex by assumption. The
is_strictly_convex() method checks whether or not a cone contains any
lines. Therefore to test whether or not a cone is proper, all we need
to check is whether or not it is solid and is_strictly_convex()
returns True.

This commit adds the is_solid() method, and uses it to implement
is_proper().