Trac #18454: Remove the 2*dim restriction on rays in random_cone().
Restricting the number of rays to twice the dimension of the lattice
for a random cone is not a natural restriction. It was intended to
prevent an infinite loop, but we can actually special-case the
problematic dimensions.

This commit removes the 2*dim upper limit and adjusts the
documentation. In exchange, it adds the restriction that in dimensions
zero, one, and two, you may not request more than zero, two, or four
rays respectively.

The random cone generation algorithm has also been updated to throw
out the cone if a search for more rays becomes futile. This makes it
possible for us to find a cone (for example) in three dimensions with
seven rays. This behavior is tested.

The generation algorithm is now described in an ALGORITHM block within
the docstring.