Trac #18704: Fix two Sphinx cross-references in geometry/cone.py.
Building the documentation with --warn-links revealed two errors in
geometry/cone.py. The first was a reference to the Python built-in
`frozenset`: apparently we have no way to link to builtins, so the
cross reference was removed. The second error referred to an obsolete
`ray_basis()` method; the surrounding docs have been updated to point
to existing methods.