Trying to figure out getting rid of sign and matrix
I tried including `del sign` and `del matrix` but this caused the tests
to fail. This commit changes the problem to just have `sage` display
which is also not ideal. I think the problem could be due to the lazy
import so am going to create another branch and see if I can fix the
problem.