Trac #18696: Add lineality() method for convex cones.
The lineality of a cone is defined to be the dimension of the largest
subspace contained within that cone. It comes up in convex analysis,
and is related to the strict convexity (or pointedness) of the cone as
well as the dimension of its dual. This commit adds the new method and
some examples/tests.