Trac #18701: Add a codim() method for convex cones.
The dimension of a cone is defined to be the dimension of the span of
its generators. It then makes sense to define the codimension as the
dimension of the orthogonal complement of that span. This number is
equal to the lineality (see trac #18696) of the dual cone.

This commit adds the new method, some examples, and a few random test
cases (based on trac #18454). The new method is added to the superclass
IntegralRayCollection since it is implemented using only `dim()` and
`lattice_dim()` from that superclass.