Fix issue in PolynomialSequence()
The constructor of PolynomialSequence treats coefficien rings of
characteristic 2 specially when really it should treat finite fields
of characteristic 2 specially. As a result

PolynomialSequence(R, [])

will fail when R.characteristic() == 2 but R.degree() is not defined.
This happens in particular for infinite fields of characteristic 2.

The fix is to check whether R is a finite field and treat it generically
if not.