Trac #19201: FiniteStateMachine.disjoint_union: fix doctests
In FiniteStateMachine.disjoint_union, doctests had Automata with
output letters.  Remove those output letters.

Furthermore, fix the labels of the states in the final example in
order to be more robust with respect to random order of the states due
to sets used in determinisation. Otherwise, a minor modification in
#19199 would lead to other labels and breaking the doctest.